// src/models/AdvisorAgreementData.ts
export default interface AdvisorAgreementData {
    partyName: string;
    partyEmail: string;
    partyAddress: string;
    partyWallet: string;
    date: string;
    counterPartyName: string;
    counterPartyEmail: string;
    counterPartyAddress: string;
    counterPartyWallet: string;
    state: string;
    typeOfCompany: string;
    stockPlanName: string;
    percentageVest: string;
    anniversaryMonth: string;
    purchaseOption: string;
    numberOfShares: string;
    stockPlanNameValue: string;
    termsConditions: string;
    vestingCommencement: string;
    typeOfTriggerAcceleration: string;
    typeOfPrice: string;
    acceptionOption: string;
    numberOfYears: string;
};